Rework jooq codegen for consistent clean builds
