+++ build-logic/src/main/kotlin/constants.kt
@@ -1,3 +1,5 @@
 import org.gradle.api.JavaVersion
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 
 val JAVA_TARGET = JavaVersion.VERSION_17
+val JVM_TARGET = JvmTarget.JVM_17
