+++ gradle/libsServer.versions.toml
@@ -9,14 +9,15 @@ logback = "1.5.6"
 koinKtor = "4.0.0"
 icu4j = "74.2"
 mockk = "1.13.10"
+jooqGradle = "9.0"
 
 [plugins]
 shadowjar = { id = "com.github.johnrengelman.shadow", version = "8.1.1" }
-jooq = { id = "nu.studer.jooq", version = "9.0" }
-flyway = { id = "org.flywaydb.flyway", version = "10.1.0" }
+jooq = { id = "nu.studer.jooq", version.ref = "jooqGradle" }
+flyway = { id = "org.flywaydb.flyway", version = "10.19.0" }
 
 [libraries]
-flyway-core = { module = "org.flywaydb:flyway-core", version = "10.1.0" }
+flyway-core = { module = "org.flywaydb:flyway-core", version = "10.19.0" }
 fastObjectPool = { module = "cn.danielw:fast-object-pool", version = "2.2.1" }
 mockk = { module = "io.mockk:mockk", version.ref = "mockk" }
 
@@ -53,6 +54,7 @@ jooq-kotlin = { module = "org.jooq:jooq-kotlin", version.ref = "jooq" }
 jooq-coroutines = { module = "org.jooq:jooq-kotlin-coroutines", version.ref = "jooq" }
 jooq-test = { module = "org.jooq:jooq-test", version.ref = "jooq" }
 jooq-test-kotlin = { module = "org.jooq:jooq-test-kotlin", version.ref = "jooq" }
+jooq-gradle = { module = "nu.studer:gradle-jooq-plugin", version.ref = "jooqGradle" }
 
 [bundles]
 icu4j = ["icu4j", "icu4j-charset", "icu4j-localespi"]
